Remove usage of Scopes type from auth module
We do not want to validate the scopes strings in the auth module, but rather just pass them through. Therefore, a Scopes type makes no sense, and we will now just pass them around as a set of strings.
If they are incorrect or faulty, it is the backend's job to report back.