Turn on automatic releases on MavenCentral
This option enables MavenCentral to automatically publish new releases we push.
Since all we do there is clicking a button anyways, this is a good option to activate for us.
It will also make sure that engineers who haven't set up Sonatype access can successfully release,
which reduces the risk of trouble if no one can be found to help out.