Provide DefaultEventReporter with a TestScope in the PlayLog test
This means refactoring DI to inject a TestScope for DefaultEventReporter so we
can perform assertions on what we send to EventProducer reliably. Some of the
new symbols are introduced in other modules, so they have been prefixed to
indicate that they are there because of the PlayLogTest. Other renames have
ocurred for consistency with this.

The new DI setup allows providing DefaultEventReporter with a TestScope,
meaning that  the for-now only test now runs inside a runTest block, and its
wait for MediaProductEnded switches out the CoroutineContext away from the test
one into a regular one so that the wait actually happens since we need playback
to occur for real.