Remove RepeatableFlakyTest
These tests are no longer flaky when it comes to asserting before reporting has
finished, so the repeat mechanism is no longer required in that sense.

If the timeout for playback proves to be a problem, then that should be bumped
instead, as it is virtually higher for other tests (they use the same timeout
for playback completion, even though said playback is not full).