TAP-603: Improve Mutex usage to fix the concurrency issue
Currently if multiple threads enter "getCredentials()" at the same time when the token is expired, they both will fire refresh logic. Because each of them get their own copy of "Tokens" object (using "getLatestTokens()" method) and then they all pass the same outdated value into the if statements that are below it. To fix this we need to put mutex a bit higher in the logic and put "getCredentials()" logic inside the mutex block, this way we ensure that whoever enter the block will get the most recent state of our "Tokens"