Use unix time stamps instead of Instant
kotlinx.datetime under the hood calls some Java APIs not available on Android API 25 and lower. This forces clients using this library to use desugaring.
We don't wan't that. And since time calculations are not very crucial in this project, we just use seconds since epoch instead, which can just be a Long.