Temporarily add kotlinx.datetime to eventproducer deps
This is due to the fact that this dependency was provided via 'api' from auth, which is a mistake.
Because of that eventproducer no needs the lib to resolve the 'Instant' type, since it uses the released binary of auth which still has it.

This will be addressed in a subsequent PR once we've released the lates auth changes.