Create LegacyCredentialsMigrator to help loading old credentials versions
We've changed the Credentials type twice, so there are two legacy formats of that type that we possibly encounter when loading data from EncryptedSharedPreferences.
This class and the legacy types takes care of trying and deserializing them in a clean way, so we don't have to pollute our DefaultTokenStore with the necessary logic.

This also means reintroducing the kotlinx.datetime dependency - however, this won't lead to any issues with 3rd party apps, and our own one is protected since it's using desugaring.