Rework authorization handling for OkHttpClients
Currently, we use our AuthorizationInterceptor on everything, but have an allow list to limit where the auth header is actually applied. This list can quickly get out of date(which it already has) and would need updates from time to time. It also featured a nasty bug where the Authenticator didn't account for this allow list, effectively adding auth header to all failed requests that would hit here. A better approach is to use different OkHttpClients for different purposes and let that steer the use of setting the auth header instead.

We know in which paths we need to put the header on, so this is a much more suitable place to do it, rather than relying on host endpoints, which can change on backend without us knowing it.

Here we basically go back to what we had prior to our HlsMediaSourceFactory added support for playing from a URL and instead create a separate path for this where we inject a new OkHttpClient that uses the AuthorizationInterceptor, and the other one doesn't anymore.