Run ktlint -F after generation to clean up imports
openapi-generator has a poat-processiong option, but since we're using a dedicated script to run ktlint, I decided to add the step for this here.

With this change, ktlint will try and autoformat as many violations as possible. While not all of them get fixed, especially cleaning up imports is quite useful, as the generator creates a lot of unnecessary import statements.